safari-600.7-branch/LayoutTests/inspector-protocol/dom/dom-search.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/protocol-test.js"></script>
<script>
function test()
{
    // Loading the queries from external file to avoid having them show up in the results.
    InspectorTest.importScript("../../../../inspector-protocol/dom/resources/dom-search-queries.js");

    // Create a DOM listener to convert nodeIds to tag names.
    InspectorTest.importScript("InspectorDOMListener.js");
    var dom = createDOMListener();

    // Caching the output to avoid searching through the log.
    var output = [];

    InspectorTest.sendCommand("DOM.getDocument", {}, onGotDocument);

    function onGotDocument(message) {
        InspectorTest.checkForError(message);
        dom.collectNode(message.result.root);
        performSearches(domSearchQueries, testFinished);
    }

    function performSearches(list, callback)
    {
        function next() {
            if (list.length)
                search(list.shift(), next);
            else
                callback();
        }
        next();
    }

    function search(query, callback)
    {
        output.push("=== Query: " + JSON.stringify(query) + " ===");
        InspectorTest.sendCommand("DOM.performSearch", {query: query}, function(message) {
            InspectorTest.checkForError(message);
            printSearchResults(message.result, callback);
        });
    }

    function printSearchResults(results, callback)
    {
        output.push("Count: " + results.resultCount);
        if (!results.resultCount)
            return callback();

        var options = {"searchId": results.searchId, "fromIndex": 0, "toIndex": results.resultCount};
        InspectorTest.sendCommand("DOM.getSearchResults", options, function onResultsReceived(message) {
            for (var nodeId of message.result.nodeIds)
                output.push(dom.getNodeIdentifier(nodeId));
            callback();
        });
    }

    function testFinished()
    {
        InspectorTest.log(output.join("\n"));
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()" class="main-frame">
    <p>Testing DOM.performSearch with no parent node ids.</p>
    <iframe src="resources/dom-search-iframe.html"></iframe>
</body>
</html>
